5-multi-agent/requirements.txt
google-adk==0.1.0
yfinance
